Bump psych version to 3.1.0
This fixes this error:
```
Your bundle is locked to psych (3.0.0), but that version could not be
found in any of the sources listed in your Gemfile. If you haven't
changed sources, that means the author of psych (3.0.0) has removed it.
You'll need to update your bundle to a version other than psych (3.0.0)
that hasn't been removed in order to install.
```
which happened because psych 3.0.0 was yanked from rubygems:
https://rubygems.org/gems/psych/versions