Use a configurable inflector via `xml_inflector`
This uses Dry::Core::Inflector's inflector detection, which will use ActiveSupoprt's inflector, if it is present, otherwise will fall back to dry-inflector (which becomes a gem dependency, to ensure that at least one inflector is available.

With this change, we adjust the `xml_convention` behavior when it is provided a symbol. If that symbol is a public method on the inflector, then that method is turned into  the naming contention proc. Otherwise, we fall back to the existing behaviour (calling .to_proc on the provided value).